+++ src/Pages/Home/components/GameSlider/GameSlider.jsx
