+++ src/Pages/Login/Login.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 const Login = () => {
     return (
         <div>
-            
+           
         </div>
     );
 };
