+++ src/Routes/Routes/Routes.js
@@ -1,6 +1,6 @@
 import { createBrowserRouter } from "react-router-dom";
 import Main from "../../Layout/Main/Main";
-import Home from "../../Pages/Home/Home/Home";
+import Home from "../../Pages/Home/Home";
 
  const router = createBrowserRouter([
     {
